# Server/swagger.yaml
swagger: '2.0'
info:
  title: Domain Collecting Tool API
  version: "2.0.0"
# the domain of the service
host: localhost:8080
# array of all schemes that your API supports
schemes:
  - http
# will be prefixed to all paths
basePath: /api

paths:

  /public/login:
    post:
      summary: Try to login
      tags:
        - public
      parameters:
        - name: u
          in: query
          description: ID of `country` option.
          required: true
          type: string
          format: string
        - name: p
          in: query
          description: Literal value of `site` option.
          required: true
          type: string
          format: string
      responses:
        302:
          description: Will redirect to `/` if success, otherwise to `/login`.
        500:
          description: Something is wrong.
          
  /public/logout:
    get:
      summary: Delete session and logout
      tags:
        - public
      responses:
        200:
          description: Successfully logged out
        500:
          description: Something is wrong
          
  /public/domain:
    put:
      summary: Get category list
      tags:
        - public
      responses:
        200:
          description: Success
        500:
          description: Fail
    get:
      summary: Check if a domain is existed.
      tags:
        - public
      responses:
        200:
          description: Existed
        404:
          description: Not Found
        500:
          description: Something is wrong
          
  /public/site:
    get:
      summary: Get category list
      tags:
        - public
      parameters:
        - name: country
          in: query
          description: ID of `country` option.
          required: true
          type: number
          format: integer
      responses:
        200:
          description: An object
          schema:
            type: object
            properties:
               sid:
                  type: number
                  description: site id.
               cid:
                  type: number
                  description: category id.
               country:
                  type: string
                  description: country name.
               site:
                  type: string
                  description: site name.
                  
  /public/csv:
    get:
      summary: Get category list
      tags:
        - public
      responses:
        200:
          description: An object
          schema:
            type: object
            properties:
               success:
                  type: boolean
                  description: .
               sid:
                  type: number
                  description: site id.
               cid:
                  type: number
                  description: category id.
                  
  /public/xslx:
    post:
      summary: download a xlsx file with provided params. Eg. `[{"sid":1,"site":"krcepp_lenovo_unit","domain":"qq.ccom","status":"2016-12-07T05:18:03.000Z","applicant":"user"},{"sid":2,"site":"kreducation_unit","domain":"163.com","status":"2016-12-12T05:18:18.000Z","applicant":"admin"}]`
      tags:
        - public
      parameters:
        - name: 'null'
          in: query
          description: json retrieved from the database
          required: true
          type: string
          format: json
      responses:
        200:
          description: file
          
  /admin/domain:
    get:
      summary: Get category list
      tags:
        - public
      responses:
        200:
          description: An object
          schema:
            type: object
            properties:
               success:
                  type: boolean
                  description: .
               sid:
                  type: number
                  description: site id.
               cid:
                  type: number
                  description: category id.
  /admin/country:
    put:
      summary: Add a site with relationship
      parameters:
        - name: country
          in: query
          description: ID of `country` option.
          required: true
          type: number
          format: integer
        - name: site
          in: query
          description: Literal value of `site` option.
          required: true
          type: string
          format: string
      tags:
        - admin
      responses:
        200:
          description: An object
          schema:
            type: object
            properties:
               success:
                  type: boolean
                  description: .
               sid:
                  type: number
                  description: site id.
               cid:
                  type: number
                  description: category id.
    delete:
      summary: Add a site with relationship
      description: |
        
      parameters:
        - name: site
          in: query
          description: ID of `site` option.
          required: true
          type: number
          format: integer
      tags:
        - admin
      responses:
        200:
          description: Success
        500:
          description: Error
    post:
      summary: Add a site with relationship
      description: |
        
      parameters:
        - name: site
          in: query
          description: ID of `site` option.
          required: true
          type: number
          format: integer
      tags:
        - admin
      responses:
        200:
          description: Success
        500:
          description: Error
  /admin/site:
    put:
      summary: Add a site with relationship
      parameters:
        - name: country
          in: query
          description: ID of `country` option.
          required: true
          type: number
          format: integer
        - name: site
          in: query
          description: Literal value of `site` option.
          required: true
          type: string
          format: string
      tags:
        - admin
      responses:
        200:
          description: An object
          schema:
            type: object
            properties:
               success:
                  type: boolean
                  description: .
               sid:
                  type: number
                  description: site id.
               cid:
                  type: number
                  description: category id.
    delete:
      summary: Delete a site
      description: |
        
      parameters:
        - name: site
          in: query
          description: ID of `site` option.
          required: true
          type: number
          format: integer
      tags:
        - admin
      responses:
        200:
          description: Success
        500:
          description: Error
    post:
      summary: Add a site with relationship
      description: |
        
      parameters:
        - name: site
          in: query
          description: ID of `site` option.
          required: true
          type: number
          format: integer
      tags:
        - admin
      responses:
        200:
          description: Success
        500:
          description: Error
          
  /admin/user:
    put:
      summary: Add a user
      parameters:
        - name: name
          in: query
          description: Literal value of `name` option.
          required: true
          type: string
          format: string
        - name: password
          in: query
          description: Literal value of `password` option.
          required: true
          type: string
          format: string
        - name: role
          in: query
          description: Literal value of `role` option. `9` = `admin`, `1` = user`
          required: true
          type: number
          format: integer
        - name: country
          in: query
          description: ID of `country` option.
          required: true
          type: number
          format: integer
      tags:
        - admin
      responses:
        200:
          description: Success.
        304:
          description: User already exists.